+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## Unreleased
+- Makes `Response.prototype.arrayBuffer` standard-compliant by returning `ArrayBuffer`.
+
 ## 0.1.337 (Dec 18, 2015)
 - Private release for profit and pleasure.